clean up cart route and data
